some changes for config
added main_path from config and search_empty_folders in editor/sort
added only_h in scripts/toc_from_titles
added try-except(eof) in merge_books/main for input

+++ epubeditor/editor/sort.py
@@ -6,6 +6,19 @@ from prompt_toolkit.completion import PathCompleter
 from epubeditor.metadata_editor.get_metadata import getMetadata
 from epubeditor.editor.book_renamer import getRoot
 from epubeditor.prompt_input import input
+from epubeditor import config
+
+main_path = None
+search_empty_folders = True
+
+if config:
+    if 'sort' in config:
+        if 'main_path' in config['sort']:
+            main_path = config['sort']['main_path']
+
+        if 'search_empty_folders' in config['sort']:
+            search_empty_folders = config['sort']['search_empty_folders']
+
 
 def sort(book, main_path):
     root = getRoot(book)
@@ -54,12 +67,14 @@ def main(books):
         get_paths=lambda: ['.'],
     )
     
-    # Получение пути к главной папке для сортировки
-    main_path = input('Main folder for sort', completer = path_completer)
-    if main_path[:2] == '~/':
-        main_path = Path.home() / main_path[2:]
+    if main_path is None:
+        # Получение пути к главной папке для сортировки
+        main_path = input('Main folder for sort', completer = path_completer)
     else:
-        main_path = Path(main_path).resolve()
+        if main_path[:2] == '~/':
+            main_path = Path.home() / main_path[2:]
+        else:
+            main_path = Path(main_path).resolve()
     
     while not main_path.is_dir():
         main_path = input('Not valid folder, try again: ', completer = path_completer)
@@ -76,25 +91,25 @@ def main(books):
     else:
         new_books.append(sort(books[0], main_path))
     
-    
-    # Удаление пустых папок
-    console = Console()
-    with console.status('[green]Searching empty folders...[/]'):
-        removed_any = True
-        while removed_any:
-            removed_any = False
-            empty_folders = []
-            for folder in main_path.rglob('*'):
-                if folder.is_dir() and not any(folder.iterdir()):
-                    empty_folders.append(folder)
-            
-            for empty_folder in sorted(empty_folders, key = lambda x: len(x.parts), reverse = True):
-                try:
-                    empty_folder.rmdir()
-                    removed_any = True
-                    console.log(f"Empty folder removed: {empty_folder}")
-                except OSError as e:
-                    console.log(f"Failed to remove {empty_folder}: {e}")
+    if search_empty_folders:
+        # Удаление пустых папок
+        console = Console()
+        with console.status('[green]Searching empty folders...[/]'):
+            removed_any = True
+            while removed_any:
+                removed_any = False
+                empty_folders = []
+                for folder in main_path.rglob('*'):
+                    if folder.is_dir() and not any(folder.iterdir()):
+                        empty_folders.append(folder)
+                
+                for empty_folder in sorted(empty_folders, key = lambda x: len(x.parts), reverse = True):
+                    try:
+                        empty_folder.rmdir()
+                        removed_any = True
+                        console.log(f"Empty folder removed: {empty_folder}")
+                    except OSError as e:
+                        console.log(f"Failed to remove {empty_folder}: {e}")
     
     return new_books
